reduce bloat of build app and move table building into build componenets
